Versioning as much as I could of the docket at one time.
This includes actual PDFs, and SHAs and ETags for each downloaded PDF.
Some PDFs could not be downloaded (>50 at least), meaning we're due for
alerts on those at some indeterminate time. Cron will be off for a bit,
as we attempt to give the Court some time to fix their cache errors.